Make constraints cloning more efficient
Documents constraints are implemented as trees of dictionaries, where the
leaves are lists of tuples. Parsing a query means building such trees.

In building the tree it's often necessary to clone part of it so that the
new branch can independently receive additional constraints, to do so
`copy.deepclone` was used.

With the advent of queries with many branches the tree cloning showed up
as a boottleneck. This commit replaces the deep copy with a shallow copy
where all the tree structure is cloned excpet for the tuples in the
leaves, the tuples being immutable and not requiring a deep clone.

The time saved is considerable and gives a 2.5x boost.

Before:
```
$ python3 -m unittest -v tests/test_emitter_*.py -k test_rules_collection
test_rules_collection (tests.test_emitter_rules.TestRules.test_rules_collection) ... ok

----------------------------------------------------------------------
Ran 1 test in 228.107s

OK
```

After:
```
$ python3 -m unittest -v tests/test_emitter_*.py -k test_rules_collection
test_rules_collection (tests.test_emitter_rules.TestRules.test_rules_collection) ... ok

----------------------------------------------------------------------
Ran 1 test in 91.882s

OK
```